Drop iOS 5 support. Updated deprecated methods and adjusted for new warnings

+++ Source/Headers/iPhone/CedarApplicationDelegate.h
@@ -1,7 +1,7 @@
 #import <UIKit/UIKit.h>
 
 int runSpecsWithinUIApplication();
-void exitWithStatusFromUIApplication(int status);
+void exitWithStatusFromUIApplication(NSInteger status);
 
 @class CDRExampleReporterViewController;
